added mixpanel optional logging to the api calls
