+++ app/common/response_templates.py
@@ -2,7 +2,7 @@ import json
 import time
 from datetime import datetime
 
-from app.common.datadog_signals import LogApiCallWeight
+from app.common.signals import LogApiCallWeight
 from app.common.datatypes import DataTypes
 from app.common.rate_limit import increment_call_rate, RateLimiter, ceil_dt_to_future_time
 from app.common.scoring_conf import ScoringMethods
